\name{learn_barchart}
\alias{learn_barchart}

\title{Bar Chart Tutorital

}
\description{learn_barchart is a tutorial on how to build a bar chart from the ggplot2 library. The tutorial uses a dataframe named 'practice' for visualization.

}
\usage{where x is a value between 0-5.
learn_barchart(x)

}

\author{Sahasrabudhe, Mayrer, Patel, Jacobs

}

\seealso{squirrel_ggplot2() learn_boxplot() learn_scatterplot() learn_piechart() learn_histogram()

\examples{For a short description of the uses of a bar chart

learn_barchart(0)

}

