\name{learn_lapply}
\alias{learn_lapply}
\title{Squirrel's lapply Tutorial
}
\description{learn_lapply is a function that provides a step by step tutorial on how to use the lapply function from the apply family of functions on a basic list with some basic functions.
}
\details{Table of contents

learn_lapply(0): Introduction

learn_lapply(1): List Creation

learn_lapply(2): List Printing

learn_lapply(3): Basic R Functions

learn_lapply(4): Model Diagnostics

learn_lapply(5): Function Creation

learn_lapply(6): Final Notes
}
\usage{Where x is a value between 0-6
learn_lapply(x)
learn_lapply(0) is the first module of the tutorial.
learn_lapply(6) is the last module of the tutorial.
}
\author{Sahasrabudhe, Mayrer, Patel, Jacobs
}
\seealso{squirrel()' to go back to the beginning.

'learn_apply(0)' to learn about the sapply function.

'learn_tapply(0)' to learn about the lapply function.

'learn_sapply(0)' to learn about the tapply function.

'squirrel_ggplot2()' to learn about data visualization using ggplot2.
}
\examples{For a short description of lapply
enter learn_lapply(0)
}