\name{learn_scatterplot}
\alias{learn_scatterplot}
\title{Squirrel's Scatterplot Tutorial
}
\description{learn_scatterplot is a function that provides a step-by-step tutorial on how to build a scatterplot from the ggplot2 library. The dataset 'iris' is utilized for visualization. This tutorial is part of the squirrel() package.
}
\details{Table of Contents

learn_lapply(0): Introduction

learn_lapply(1): Data Loading

learn_lapply(2): Data and Variable Information

learn_lapply(3): The Scatterplot Geom

learn_lapply(4): Changing the Plots Points

learn_lapply(5): ggplot2 Labeling

learn_lapply(6): Applying a theme

learn_lapply(7): Final Thoughts

}
\usage{Where x is a value between 0-7
learn_scatterplot(x)
learn_scatterplot(0) is the first module of the tutorial.
learn_scatterplot(7) is the last module of the tutorial.
}
\author{Sahasrabudhe, Mayrer, Patel, Jacobs
}
\seealso{squirrel()' to go back to the beginning.

'learn_apply(0)' to learn about the sapply function.

'learn_tapply(0)' to learn about the lapply function.

'learn_sapply(0)' to learn about the tapply function.

'squirrel_ggplot2()' to learn about data visualization using ggplot2.
}
\examples{For a short description of a scatterplot
enter learn_scatterplot(0)
}